refactor: rename Page to PageButton
